Change deposit method in Vending Machine to actually buy the product.
When enough coins are inserted Vending Machine will ask ATM if returning
change is possible.
If yes, Vending Machine will deposit the money and dispose the
product.
If not, Vending Machine will print a warning and return money to
the user.

Fix error in dispense algorithm - improper number of used coins
was returned.

Change ZERO created in VendingMachine class to correct BigDecimal.ZERO.